+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -4,6 +4,7 @@ import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import java.util.List;
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public final class HubAdapter implements IHub {
@@ -177,6 +178,16 @@ public final class HubAdapter implements IHub {
   @Override
   public @Nullable SentryTraceHeader traceHeaders() {
     return Sentry.traceHeaders();
+  }
+
+  @Override
+  public void setSpanContext(final @NotNull Throwable t, final @NotNull SpanContext sc) {
+    Sentry.getCurrentHub().setSpanContext(t, sc);
+  }
+
+  @Override
+  public @Nullable SpanContext getSpanContext(final @NotNull Throwable ex) {
+    return Sentry.getCurrentHub().getSpanContext(ex);
   }
 
   @Override